CORE-12000 - Consolidate commands from mgm and network plugin (#4176)
This task is about consolidating commands available in both into a single plugin. Proposal is to retain the name network, which will be a better fit for hosting capabilities related to application networks (not necessarily only about MGM).
