+++ settings.gradle
@@ -445,7 +445,6 @@ include 'tools:plugins:plugins-rest'
 include 'tools:plugins:db-config'
 include 'tools:plugins:initial-config'
 include 'tools:plugins:initial-rbac'
-include 'tools:plugins:mgm'
 include 'tools:plugins:network'
 include 'tools:plugins:package'
 include 'tools:plugins:preinstall'
