--- tools/plugins/mgm/build.gradle
@@ -1,56 +1,0 @@
-plugins {
-    id 'distribution'
-    id 'org.jetbrains.kotlin.jvm'
-    id 'org.jetbrains.kotlin.kapt'
-    id 'corda.cli-plugin-packager'
-    id 'corda.common-publishing'
-}
-
-ext {
-    releasable = false
-}
-
-group 'net.corda.cli.deployment'
-
-dependencies {
-    compileOnly "net.corda.cli.host:api:$pluginHostVersion"
-
-    implementation "com.fasterxml.jackson.module:jackson-module-kotlin:$jacksonVersion"
-    implementation "org.yaml:snakeyaml:${snakeyamlVersion}"
-    implementation "com.konghq:unirest-java:$unirestVersion"
-    implementation project(":libs:crypto:certificate-generation")
-    implementation project(':libs:crypto:cipher-suite')
-    implementation "org.pf4j:pf4j:${pf4jVersion}"
-    kapt "org.pf4j:pf4j:${pf4jVersion}"
-
-    implementation project(":tools:plugins:package")
-    implementation project(':tools:plugins:plugins-rest')
-    implementation project(':components:membership:membership-rest')
-
-    testImplementation "net.corda.cli.host:api:$pluginHostVersion"
-    testImplementation "org.pf4j:pf4j:${pf4jVersion}"
-    testImplementation "org.junit.jupiter:junit-jupiter:${junit5Version}"
-    testImplementation "com.github.stefanbirkner:system-lambda:1.2.1"
-}
-
-cliPlugin {
-    cliPluginClass = 'net.corda.cli.plugins.mgm.MgmPluginWrapper'
-    cliPluginDescription = 'Plugin for membership operations.'
-}
-
-tasks.named("installDist") {
-    dependsOn cliPlugin
-    def homePath = System.properties['user.home']
-    from cliPlugin
-    into "$homePath/.corda/cli/plugins"
-}
-
-publishing {
-    publications {
-        maven(MavenPublication) {
-            artifactId "${cliPlugin.cliPluginId.get()}-cli-plugin"
-            groupId project.group
-            artifact cliPluginTask
-        }
-    }
-}
