+++ tools/plugins/network/build.gradle
@@ -15,25 +15,27 @@ group 'net.corda.cli.deployment'
 dependencies {
     compileOnly "net.corda.cli.host:api:$pluginHostVersion"
 
-    implementation project(":tools:plugins:plugins-rest")
-    implementation project(":components:membership:membership-rest")
-
     implementation "com.fasterxml.jackson.module:jackson-module-kotlin:$jacksonVersion"
     implementation "org.yaml:snakeyaml:${snakeyamlVersion}"
+    implementation "com.konghq:unirest-java:$unirestVersion"
+    implementation project(":libs:crypto:certificate-generation")
+    implementation project(':libs:crypto:cipher-suite')
     implementation "org.pf4j:pf4j:${pf4jVersion}"
     kapt "org.pf4j:pf4j:${pf4jVersion}"
+
+    implementation project(":tools:plugins:package")
+    implementation project(':tools:plugins:plugins-rest')
+    implementation project(':components:membership:membership-rest')
 
     testImplementation "net.corda.cli.host:api:$pluginHostVersion"
     testImplementation "org.pf4j:pf4j:${pf4jVersion}"
     testImplementation "org.junit.jupiter:junit-jupiter:${junit5Version}"
-    testImplementation "org.mockito:mockito-core:$mockitoVersion"
-    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
     testImplementation "com.github.stefanbirkner:system-lambda:1.2.1"
 }
 
 cliPlugin {
     cliPluginClass = 'net.corda.cli.plugins.network.NetworkPluginWrapper'
-    cliPluginDescription = 'Plugin for interacting with a network.'
+    cliPluginDescription = 'Plugin for network operations.'
 }
 
 tasks.named("installDist") {
