+++ tools/plugins/network/src/main/kotlin/net/corda/cli/plugins/network/Helpers.kt
@@ -1,4 +1,4 @@
-package net.corda.cli.plugins.mgm
+package net.corda.cli.plugins.network
 
 import java.net.ServerSocket
 import kotlin.concurrent.thread
