+++ tools/plugins/network/src/main/resources/certificates/certificate1.pem
