+++ tools/plugins/network/src/main/resources/certificates/certificate2.pem
