+++ tools/plugins/network/src/main/resources/certificates/gradle-plugin-default-key.pem
